rename: button 컴포넌트 위치 변경
