+++ components/detail/Rooms/Index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { roomList } from '@/data/detail/roomData';
 import { useRouter } from 'next/router';
 import { formatNumberWithCommas } from 'utils/formatNumberWithCommas';
-import { BlackButton, WhiteButton } from '../Top/global/button';
+import { BlackButton, WhiteButton } from '../../global/button/BasicButton';
 
 type Props = {};
